requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-fernet-fields==0.6
django-rest-registration==0.5.6
djangorestframework==3.11.1
httplib2==0.18.1
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
typing==3.7.4.3
urllib3==1.25.10
